SP3 -> Task One - > Part Two and Three

+++ SP/SP3/solution-jpa-mapping-exercise-three/src/main/java/domain/DiscountType.java
@@ -38,7 +38,7 @@ public class DiscountType {
         this.customerList = customerList;
     }
 
-    public double calcDiscount(int pricePerItem, int quantity) {
+    public double calcDiscount(double pricePerItem, int quantity) {
         return pricePerItem*quantity;
     }
 }
